Implemented Cache
Cache will be checked before looping through the URL.  If it is found, the correct iteration through URL will be moved to or it will be skipped if the rec_type is "A".  When getting a new server, it will be cached along with what the URL was and the type of record.